+++ lib/deployment.sh
@@ -16,10 +16,10 @@ function preDeploy() {
     trace "Status looks good"
   else
     # If running in --force mode we will not allow deployment to continue
-    if [[ "${FORCE}" == "1" ]] && [[ "${APPROVE}" != "1" ]]; then
+    if [[ "${FORCE}" == "1" && "${APPROVE}" != "1" ]] || [[ "${UPDATEONLY}" == "1" ]]; then
       trace "Checking for files that need stashing"
       # Stash the dirty bits
-      if [[ "${STASH}" == "TRUE" ]]; then
+      if [[ "${STASH}" == "TRUE" ]] || [[ "${UPDATEONLY}" == "1" ]]; then # Bah, clunky
         emptyLine
         trace "Stashing dirty files"
         if [[ "${VERBOSE}" == "1" ]] && [[ "${QUIET}" != "1" ]]; then