+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Added option for email notification to be sent "clean," without full logging information
 ### Changed
 - Running as root is no longer allowed; this can be overridden by setting `ALLOWROOT="TRUE"`
+- Slightly improved error checking
 
 ## [3.6] - 07-08-2017
 ### Added